+++ odps-sdk/odps-sdk-core/src/main/java/com/aliyun/odps/utils/ExceptionUtils.java
@@ -1,5 +1,7 @@
 package com.aliyun.odps.utils;
 
+import java.util.Collection;
+
 public class ExceptionUtils {
 
     public static void checkStringArgumentNotNull(String arg, String value) {
@@ -7,6 +9,12 @@ public class ExceptionUtils {
             throw new IllegalArgumentException("Argument '" + arg + "' cannot be null or empty");
         }
     }
+    public static void checkCollectionArgumentNotNull(String arg, Collection value) {
+        if (value == null || value.isEmpty()) {
+            throw new IllegalArgumentException("Argument '" + arg + "' cannot be null");
+        }
+    }
+
     public static void checkArgumentNotNull(String arg, Object value) {
         if (value == null) {
             throw new IllegalArgumentException("Argument '" + arg + "' cannot be null");